+++ src/argouml-app/src/org/argouml/persistence/UmlFilePersister.java
@@ -1,6 +1,6 @@
 /* $Id$
  *****************************************************************************
- * Copyright (c) 2009 Contributors - see below
+ * Copyright (c) 2009-2012 Contributors - see below
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,6 +9,7 @@
  * Contributors:
  *    Bob Tarling
  *    Thomas Neustupny
+ *    Michiel van der Wulp
  *****************************************************************************
  *
  * Some portions of this file was previously release using the BSD License:
@@ -137,8 +138,7 @@ public class UmlFilePersister extends AbstractFilePersister {
     }
 
     /**
-     * It is being considered to save out individual xmi's from individuals
-     * diagrams to make it easier to modularize the output of Argo.
+     * Save the project in ".uml" format.
      * 
      * @param file The file to write.
      * @param project the project to save
@@ -151,6 +151,10 @@ public class UmlFilePersister extends AbstractFilePersister {
     public void doSave(Project project, File file) throws SaveException,
         InterruptedException {
 
+        /* Retain the previous project file even when the save operation 
+         * crashes in the middle. Also create a backup file after saving. */
+        boolean doSafeSaves = useSafeSaves();
+
         ProgressMgr progressMgr = new ProgressMgr();
         progressMgr.setNumberOfPhases(4);
         progressMgr.nextPhase();
@@ -158,14 +162,16 @@ public class UmlFilePersister extends AbstractFilePersister {
         File lastArchiveFile = new File(file.getAbsolutePath() + "~");
         File tempFile = null;
 
-        try {
-            tempFile = createTempFile(file);
-        } catch (FileNotFoundException e) {
-            throw new SaveException(
-                    "Failed to archive the previous file version", e);
-        } catch (IOException e) {
-            throw new SaveException(
-                    "Failed to archive the previous file version", e);
+        if (doSafeSaves) {
+            try {
+                tempFile = createTempFile(file);
+            } catch (FileNotFoundException e) {
+                throw new SaveException(Translator.localize(
+                        "optionpane.save-project-exception-cause1"), e);
+            } catch (IOException e) {
+                throw new SaveException(Translator.localize(
+                        "optionpane.save-project-exception-cause2"), e);
+            }
         }
 
         try {
@@ -186,17 +192,19 @@ public class UmlFilePersister extends AbstractFilePersister {
                 LOG.info("Dir ==" + path);
             }
 
-            // if save did not raise an exception
-            // and name+"#" exists move name+"#" to name+"~"
-            // this is the correct backup file
-            if (lastArchiveFile.exists()) {
-                lastArchiveFile.delete();
-            }
-            if (tempFile.exists() && !lastArchiveFile.exists()) {
-                tempFile.renameTo(lastArchiveFile);
-            }
-            if (tempFile.exists()) {
-                tempFile.delete();
+            if (doSafeSaves) {
+                // if save did not raise an exception
+                // and name+"#" exists move name+"#" to name+"~"
+                // this is the correct backup file
+                if (lastArchiveFile.exists()) {
+                    lastArchiveFile.delete();
+                }
+                if (tempFile.exists() && !lastArchiveFile.exists()) {
+                    tempFile.renameTo(lastArchiveFile);
+                }
+                if (tempFile.exists()) {
+                    tempFile.delete();
+                }
             }
 
             progressMgr.nextPhase();
@@ -204,11 +212,13 @@ public class UmlFilePersister extends AbstractFilePersister {
         } catch (Exception e) {
             LOG.error("Exception occured during save attempt", e);
 
-            // frank: in case of exception
-            // delete name and mv name+"#" back to name if name+"#" exists
-            // this is the "rollback" to old file
-            file.delete();
-            tempFile.renameTo(file);
+            if (doSafeSaves) {
+                // frank: in case of exception
+                // delete name and mv name+"#" back to name if name+"#" exists
+                // this is the "rollback" to old file
+                file.delete();
+                tempFile.renameTo(file);
+            }
             if (e instanceof InterruptedException) {
                 throw (InterruptedException) e;
             } else {
