+++ src/argouml-app/src/org/argouml/persistence/XmiFilePersister.java
@@ -1,6 +1,6 @@
 /* $Id$
  *****************************************************************************
- * Copyright (c) 2009 Contributors - see below
+ * Copyright (c) 2009-2012 Contributors - see below
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,6 +8,7 @@
  *
  * Contributors:
  *    tfmorris
+ *    Michiel van der Wulp
  *****************************************************************************
  *
  * Some portions of this file was previously release using the BSD License:
@@ -117,6 +118,10 @@ class XmiFilePersister extends AbstractFilePersister
     public void doSave(Project project, File file)
         throws SaveException, InterruptedException {
 
+        /* Retain the previous project file even when the save operation 
+         * crashes in the middle. Also create a backup file after saving. */
+        boolean doSafeSaves = useSafeSaves();
+        
         ProgressMgr progressMgr = new ProgressMgr();
         progressMgr.setNumberOfPhases(4);
         progressMgr.nextPhase();
@@ -124,14 +129,16 @@ class XmiFilePersister extends AbstractFilePersister
         File lastArchiveFile = new File(file.getAbsolutePath() + "~");
         File tempFile = null;
 
-        try {
-            tempFile = createTempFile(file);
-        } catch (FileNotFoundException e) {
-            throw new SaveException(
-                    "Failed to archive the previous file version", e);
-        } catch (IOException e) {
-            throw new SaveException(
-                    "Failed to archive the previous file version", e);
+        if (doSafeSaves) {
+            try {
+                tempFile = createTempFile(file);
+            } catch (FileNotFoundException e) {
+                throw new SaveException(Translator.localize(
+                        "optionpane.save-project-exception-cause1"), e);
+            } catch (IOException e) {
+                throw new SaveException(Translator.localize(
+                        "optionpane.save-project-exception-cause2"), e);
+            }
         }
 
         OutputStream stream = null;
@@ -140,17 +147,19 @@ class XmiFilePersister extends AbstractFilePersister
             writeProject(project, stream, progressMgr);
             stream.close();
 
-            // if save did not raise an exception
-            // and name+"#" exists move name+"#" to name+"~"
-            // this is the correct backup file
-            if (lastArchiveFile.exists()) {
-                lastArchiveFile.delete();
-            }
-            if (tempFile.exists() && !lastArchiveFile.exists()) {
-                tempFile.renameTo(lastArchiveFile);
-            }
-            if (tempFile.exists()) {
-                tempFile.delete();
+            if (doSafeSaves) {
+                // if save did not raise an exception
+                // and name+"#" exists move name+"#" to name+"~"
+                // this is the correct backup file
+                if (lastArchiveFile.exists()) {
+                    lastArchiveFile.delete();
+                }
+                if (tempFile.exists() && !lastArchiveFile.exists()) {
+                    tempFile.renameTo(lastArchiveFile);
+                }
+                if (tempFile.exists()) {
+                    tempFile.delete();
+                }
             }
         } catch (InterruptedException exc) {
             try {
@@ -163,11 +172,13 @@ class XmiFilePersister extends AbstractFilePersister
                 stream.close();
             } catch (IOException ex) { }
 
-            // frank: in case of exception
-            // delete name and mv name+"#" back to name if name+"#" exists
-            // this is the "rollback" to old file
-            file.delete();
-            tempFile.renameTo(file);
+            if (doSafeSaves) {
+                // frank: in case of exception
+                // delete name and mv name+"#" back to name if name+"#" exists
+                // this is the "rollback" to old file
+                file.delete();
+                tempFile.renameTo(file);
+            }
             // we have to give a message to user and set the system to unsaved!
             throw new SaveException(e);
         }
